// src/Model/createMembershipOrPartnershipInvitesBody.ts
/**
 * Pinterest REST API
 * Pinterest\'s REST API
 *
 * The version of the OpenAPI document: 5.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InviteType } from './inviteType';


/**
 * Body to be used on path to send Members or Partners Invite or Request
 */
export interface CreateMembershipOrPartnershipInvitesBody { 
    /**
     * The business access level to grant member/partner. Note, values are case-sensitive. - EMPLOYEE: Can only view and access assets you assign them to. They cannot see details about other employees, partners, or other assets. - BIZ_ADMIN: Have full control of roles and can add employees and partners as well as grant asset access. - PARTNER: Can only view and access assets you assign them to/or they assign to you.
     */
    business_role: CreateMembershipOrPartnershipInvitesBody.BusinessRoleEnum;
    invite_type: InviteType;
    /**
     * A list of usernames, emails, or a mix of them. Should be used if invite_type is MEMBER_INVITE
     */
    members?: Array<string>;
    /**
     * A list of partner_id. Should be used if invite_type is PARTNER_INVITE or PARTNER_REQUEST
     */
    partners?: Array<string>;
}
export namespace CreateMembershipOrPartnershipInvitesBody {
    export type BusinessRoleEnum = 'EMPLOYEE' | 'BIZ_ADMIN' | 'PARTNER';
    export const BusinessRoleEnum = {
        Employee: 'EMPLOYEE' as BusinessRoleEnum,
        BizAdmin: 'BIZ_ADMIN' as BusinessRoleEnum,
        Partner: 'PARTNER' as BusinessRoleEnum
    }
}