// src/Model/inviteType.ts
/**
 * Pinterest REST API
 * Pinterest\'s REST API
 *
 * The version of the OpenAPI document: 5.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The type of invite. <br>\'MEMBER_INVITE\' is to invite a member to access your business assets. <br>\'PARTNER INVITE\' is to invite a partner to access your business assets. <br>\'PARTNER_REQUEST\' is to request access a partner\'s business assets.
 */
export type InviteType = 'MEMBER_INVITE' | 'PARTNER_INVITE' | 'PARTNER_REQUEST';

export const InviteType = {
    MemberInvite: 'MEMBER_INVITE' as InviteType,
    PartnerInvite: 'PARTNER_INVITE' as InviteType,
    PartnerRequest: 'PARTNER_REQUEST' as InviteType
}